pyproject.toml
[tool.poetry]
name = "backend-gateway"
version = "0.1.0"
description = "Django DRF gateway"
authors = ["you <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "5.0.7"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
django-cors-headers = "4.3.1"
python-dotenv = "1.0.1"
uvicorn = {version = "0.30.1", extras = ["standard"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
